fix: emergency gacha fix

+++ internal/domain/pokemon.go
@@ -4,7 +4,7 @@ type Pokemon struct {
 	ID             int
 	Name           string
 	Type           string
-	CatchRate      float64
+	CatchRate      int
 	IsRare         bool
 	RegisteredDate string
 }
